Moved Wink water heater from climate to water heater.
