+++ homeassistant/components/climate/wink.py
@@ -1,5 +1,5 @@
 """
-Support for Wink thermostats, Air Conditioners, and Water Heaters.
+Support for Wink thermostats and Air Conditioners.
 
 For more details about this platform, please refer to the documentation at
 https://home-assistant.io/components/climate.wink/
@@ -8,9 +8,9 @@ import logging
 
 from homeassistant.components.climate import (
     ATTR_CURRENT_HUMIDITY, ATTR_TARGET_TEMP_HIGH, ATTR_TARGET_TEMP_LOW,
-    ATTR_TEMPERATURE, STATE_AUTO, STATE_COOL, STATE_ECO, STATE_ELECTRIC,
-    STATE_FAN_ONLY, STATE_GAS, STATE_HEAT, STATE_HEAT_PUMP, STATE_HIGH_DEMAND,
-    STATE_PERFORMANCE, SUPPORT_AUX_HEAT, SUPPORT_AWAY_MODE, SUPPORT_FAN_MODE,
+    ATTR_TEMPERATURE, STATE_AUTO, STATE_COOL, STATE_ECO,
+    STATE_FAN_ONLY, STATE_HEAT, SUPPORT_AUX_HEAT,
+    SUPPORT_AWAY_MODE, SUPPORT_FAN_MODE,
     SUPPORT_OPERATION_MODE, SUPPORT_TARGET_TEMPERATURE,
     SUPPORT_TARGET_TEMPERATURE_HIGH, SUPPORT_TARGET_TEMPERATURE_LOW,
     ClimateDevice)
@@ -24,11 +24,9 @@ _LOGGER = logging.getLogger(__name__)
 ATTR_ECO_TARGET = 'eco_target'
 ATTR_EXTERNAL_TEMPERATURE = 'external_temperature'
 ATTR_OCCUPIED = 'occupied'
-ATTR_RHEEM_TYPE = 'rheem_type'
 ATTR_SCHEDULE_ENABLED = 'schedule_enabled'
 ATTR_SMART_TEMPERATURE = 'smart_temperature'
 ATTR_TOTAL_CONSUMPTION = 'total_consumption'
-ATTR_VACATION_MODE = 'vacation_mode'
 ATTR_HEAT_ON = 'heat_on'
 ATTR_COOL_ON = 'cool_on'
 
@@ -42,14 +40,9 @@ HA_STATE_TO_WINK = {
     STATE_AUTO: 'auto',
     STATE_COOL: 'cool_only',
     STATE_ECO: 'eco',
-    STATE_ELECTRIC: 'electric_only',
     STATE_FAN_ONLY: 'fan_only',
-    STATE_GAS: 'gas',
     STATE_HEAT: 'heat_only',
-    STATE_HEAT_PUMP: 'heat_pump',
-    STATE_HIGH_DEMAND: 'high_demand',
     STATE_OFF: 'off',
-    STATE_PERFORMANCE: 'performance',
 }
 
 WINK_STATE_TO_HA = {value: key for key, value in HA_STATE_TO_WINK.items()}
@@ -61,9 +54,6 @@ SUPPORT_FLAGS_THERMOSTAT = (
 
 SUPPORT_FLAGS_AC = (SUPPORT_TARGET_TEMPERATURE | SUPPORT_OPERATION_MODE |
                     SUPPORT_FAN_MODE)
-
-SUPPORT_FLAGS_HEATER = (SUPPORT_TARGET_TEMPERATURE | SUPPORT_OPERATION_MODE |
-                        SUPPORT_AWAY_MODE)
 
 
 def setup_platform(hass, config, add_entities, discovery_info=None):
@@ -77,10 +67,6 @@ def setup_platform(hass, config, add_entities, discovery_info=None):
         _id = climate.object_id() + climate.name()
         if _id not in hass.data[DOMAIN]['unique_ids']:
             add_entities([WinkAC(climate, hass)])
-    for water_heater in pywink.get_water_heaters():
-        _id = water_heater.object_id() + water_heater.name()
-        if _id not in hass.data[DOMAIN]['unique_ids']:
-            add_entities([WinkWaterHeater(water_heater, hass)])
 
 
 class WinkThermostat(WinkDevice, ClimateDevice):
@@ -504,93 +490,3 @@ class WinkAC(WinkDevice, ClimateDevice):
         elif fan_mode == SPEED_HIGH:
             speed = 1.0
         self.wink.set_ac_fan_speed(speed)
-
-
-class WinkWaterHeater(WinkDevice, ClimateDevice):
-    """Representation of a Wink water heater."""
-
-    @property
-    def supported_features(self):
-        """Return the list of supported features."""
-        return SUPPORT_FLAGS_HEATER
-
-    @property
-    def temperature_unit(self):
-        """Return the unit of measurement."""
-        # The Wink API always returns temp in Celsius
-        return TEMP_CELSIUS
-
-    @property
-    def device_state_attributes(self):
-        """Return the optional device state attributes."""
-        data = {}
-        data[ATTR_VACATION_MODE] = self.wink.vacation_mode_enabled()
-        data[ATTR_RHEEM_TYPE] = self.wink.rheem_type()
-
-        return data
-
-    @property
-    def current_operation(self):
-        """
-        Return current operation one of the following.
-
-        ["eco", "performance", "heat_pump",
-        "high_demand", "electric_only", "gas]
-        """
-        if not self.wink.is_on():
-            current_op = STATE_OFF
-        else:
-            current_op = WINK_STATE_TO_HA.get(self.wink.current_mode())
-            if current_op is None:
-                current_op = STATE_UNKNOWN
-        return current_op
-
-    @property
-    def operation_list(self):
-        """List of available operation modes."""
-        op_list = ['off']
-        modes = self.wink.modes()
-        for mode in modes:
-            if mode == 'aux':
-                continue
-            ha_mode = WINK_STATE_TO_HA.get(mode)
-            if ha_mode is not None:
-                op_list.append(ha_mode)
-            else:
-                error = "Invalid operation mode mapping. " + mode + \
-                    " doesn't map. Please report this."
-                _LOGGER.error(error)
-        return op_list
-
-    def set_temperature(self, **kwargs):
-        """Set new target temperature."""
-        target_temp = kwargs.get(ATTR_TEMPERATURE)
-        self.wink.set_temperature(target_temp)
-
-    def set_operation_mode(self, operation_mode):
-        """Set operation mode."""
-        op_mode_to_set = HA_STATE_TO_WINK.get(operation_mode)
-        self.wink.set_operation_mode(op_mode_to_set)
-
-    @property
-    def target_temperature(self):
-        """Return the temperature we try to reach."""
-        return self.wink.current_set_point()
-
-    def turn_away_mode_on(self):
-        """Turn away on."""
-        self.wink.set_vacation_mode(True)
-
-    def turn_away_mode_off(self):
-        """Turn away off."""
-        self.wink.set_vacation_mode(False)
-
-    @property
-    def min_temp(self):
-        """Return the minimum temperature."""
-        return self.wink.min_set_point()
-
-    @property
-    def max_temp(self):
-        """Return the maximum temperature."""
-        return self.wink.max_set_point()
